+++ tests/integration/__init__.py
@@ -12,6 +12,7 @@ import pprint
 import logging
 import tempfile
 import subprocess
+import signal
 import multiprocessing
 from hashlib import md5
 from datetime import datetime, timedelta
@@ -195,15 +196,77 @@ class TestDaemon(object):
         self.master_process = multiprocessing.Process(target=master.start)
         self.master_process.start()
 
-        minion = salt.minion.Minion(self.minion_opts)
-        self.minion_process = multiprocessing.Process(target=minion.tune_in)
-        self.minion_process.start()
+        if self.minion_opts.get('processpool', False):
+            import yaml
+            from copy import deepcopy
+            devnull = open('/dev/null' ,'w')
 
-        sub_minion = salt.minion.Minion(self.sub_minion_opts)
-        self.sub_minion_process = multiprocessing.Process(
-            target=sub_minion.tune_in
-        )
-        self.sub_minion_process.start()
+            minion_config = deepcopy(self.minion_opts)
+            minion_config['user'] = pwd.getpwuid(os.getuid()).pw_name
+            minion_config_dir =  os.path.join(
+                TMP,
+                'main_minion'
+            )
+            minion_config['pki_dir'] = os.path.join(minion_config_dir, 'pki')
+            minion_config['log_file'] = os.path.join(minion_config_dir, 'log')
+            minion_config['log_level_logfile'] = 'debug'
+            del minion_config['conf_file']
+            os.makedirs(minion_config_dir)
+            minion_config_file = os.path.join(minion_config_dir, 'minion')
+            with open(minion_config_file, 'w') as config_file:
+                config_file.write(yaml.dump(minion_config))
+            minion_cmd = [
+                os.path.join(SCRIPT_DIR, 'salt-minion'),
+                '-c',
+                minion_config_dir
+            ]
+            self.minion_process = subprocess.Popen(
+                minion_cmd,
+                stdout=devnull,
+                stderr=devnull
+            )
+        else:
+            minion = salt.minion.Minion(self.minion_opts)
+            self.minion_process = multiprocessing.Process(target=minion.tune_in)
+            self.minion_process.start()
+
+        if self.sub_minion_opts.get('processpool', False):
+            import yaml
+            from copy import deepcopy
+            devnull = open('/dev/null' ,'w')
+
+
+            sub_minion_config = deepcopy(self.sub_minion_opts)
+            sub_minion_config['user'] = pwd.getpwuid(os.getuid()).pw_name
+            sub_minion_config_dir =  os.path.join(
+                TMP,
+                'sub_minion'
+            )
+            sub_minion_config['pki_dir'] = os.path.join(sub_minion_config_dir, 'pki')
+            sub_minion_config['log_file'] = os.path.join(sub_minion_config_dir, 'log')
+            sub_minion_config['log_level_logfile'] = 'debug'
+            del sub_minion_config['conf_file']
+            os.makedirs(sub_minion_config_dir)
+            sub_minion_config_file = os.path.join(sub_minion_config_dir, 'minion')
+            with open(sub_minion_config_file, 'w') as config_file:
+                config_file.write(yaml.dump(sub_minion_config))
+
+            sub_minion_cmd = [
+                os.path.join(SCRIPT_DIR, 'salt-minion'),
+                '-c',
+                sub_minion_config_dir
+            ]
+            self.sub_minion_process = subprocess.Popen(
+                sub_minion_cmd,
+                stdout=devnull,
+                stderr=devnull
+            )
+        else:
+            sub_minion = salt.minion.Minion(self.sub_minion_opts)
+            self.sub_minion_process = multiprocessing.Process(
+                target=sub_minion.tune_in
+            )
+            self.sub_minion_process.start()
 
         smaster = salt.master.Master(self.smaster_opts)
         self.smaster_process = multiprocessing.Process(target=smaster.start)
@@ -276,10 +339,20 @@ class TestDaemon(object):
         '''
         import integration
         integration.SYNDIC = None
-        self.sub_minion_process.terminate()
-        self.sub_minion_process.join()
-        self.minion_process.terminate()
-        self.minion_process.join()
+        if self.sub_minion_opts.get('processpool', False):
+            if self.sub_minion_process.poll() is None:
+                self.sub_minion_process.terminate()
+                self.sub_minion_process.wait()
+        else:
+            self.sub_minion_process.terminate()
+            self.sub_minion_process.join()
+        if self.minion_opts.get('processpool', False):
+            if self.minion_process.poll() is None:
+                self.minion_process.terminate()
+                self.minion_process.wait()
+        else:
+            self.minion_process.terminate()
+            self.minion_process.join()
         self.master_process.terminate()
         self.master_process.join()
         self.syndic_process.terminate()
